+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- DDEV docroot moved to `.ddev/wordpress/` subdirectory to keep project root clean
+- Remove `docker-compose.plugin.yaml.dist` and `docker-compose.theme.yaml.dist` mount files
+
+### Fixed
+
+- WordPress core files polluting project root when using `ddev-orchestrate`
+
 ## [0.5.0] - 2026-04-05
 
 ### Added
@@ -74,6 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Workflow callers missing permissions (caused startup_failure)
 
+[Unreleased]: https://github.com/apermo/template-wordpress/compare/v0.5.0...HEAD
 [0.5.0]: https://github.com/apermo/template-wordpress/compare/v0.4.1...v0.5.0
 [0.4.0]: https://github.com/apermo/template-wordpress/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/apermo/template-wordpress/compare/v0.2.0...v0.3.0